Move delete logic to `EditableLink.Delete`
